+++ src/info.jl
@@ -11,6 +11,10 @@ function info_update!(
     #the pre-homogenization x and z variables.
     τinv = inv(variables.τ)
 
+    #unscaled linear term norms
+    normb = data_get_normb!(data)
+    normq = data_get_normq!(data)
+
     #shortcuts for the equilibration matrices
     d = data.equilibration.d; dinv = data.equilibration.dinv
     e = data.equilibration.e; einv = data.equilibration.einv
@@ -23,18 +27,28 @@ function info_update!(
     info.cost_primal =  (+residuals.dot_qx*τinv + xPx_τinvsq_over2)/cscale
     info.cost_dual   =  (-residuals.dot_bz*τinv - xPx_τinvsq_over2)/cscale
 
-    #primal and dual relative residuals.   Need to invert the equilibration
-    normx = scaled_norm(dinv,variables.x) * τinv
-    normz = scaled_norm(einv,variables.z) * τinv
-    norms = scaled_norm(einv,variables.s) * τinv
+    #variables norms, undoing the equilibration.  Do not unscale
+    #by τ yet because the infeasibility residuals are ratios of 
+    #terms that have no affine parts anyway
+    normx = norm_scaled(dinv,variables.x)
+    normz = norm_scaled(einv,variables.z)
+    norms = norm_scaled(einv,variables.s)
 
-    info.res_primal  = scaled_norm(einv,residuals.rz) * τinv / max(one(T), data.normb + normx + norms)
-    info.res_dual    = scaled_norm(dinv,residuals.rx) * τinv / max(one(T), data.normq + normx + normz)
+    #primal and dual infeasibility residuals.   
+    info.res_primal_inf = norm_scaled(dinv,residuals.rx_inf) / max(one(T), normz)
+    info.res_dual_inf   = max(
+        norm_scaled(dinv,residuals.Px) / max(one(T), normx),
+        norm_scaled(einv,residuals.rz_inf) / max(one(T), normx + norms)
+    )
 
-    #primal and dual infeasibility residuals.   Need to invert the equilibration
-    info.res_primal_inf = scaled_norm(dinv,residuals.rx_inf) / max(one(T), normz)
-    info.res_dual_inf   = max(scaled_norm(dinv,residuals.Px) / max(one(T), normx),
-                              scaled_norm(einv,residuals.rz_inf) / max(one(T), normx + norms))
+    #now back out the τ scaling so we can normalize the unscaled primal / dual errors 
+    normx *= τinv
+    normz *= τinv
+    norms *= τinv
+
+    #primal and dual relative residuals.  
+    info.res_primal  = norm_scaled(einv,residuals.rz) * τinv / max(one(T), normb + normx + norms)
+    info.res_dual    = norm_scaled(dinv,residuals.rx) * τinv / max(one(T), normq + normx + normz)
 
     #absolute and relative gaps
     info.gap_abs = abs(info.cost_primal - info.cost_dual)