+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Version numbering in this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).  We aim to keep the core solver functionality and minor releases in sync between the Rust/Python and Julia implementations.   Small fixes that affect one implementation only may result in the patch release versions differing.
 
+## [0.7.0] - 2024-26-02
+### Changed 
 
+- Solution output reports dual objective values.  Infeasible problems report NaN. 
+- Solver now supports problems with a mix of PSD and nonsymmetric cones. 
+- Added methods for updating problem data without reallocating memory.  
+- Bug fix enforcing scaling limits in equilibration.  
+- Bug fix in infeasibility detection. 
+
+See [Rust version changelog](https://github.com/oxfordcontrol/Clarabel.rs/compare/v0.6.0...v0.7.0) for related issues.
 
 ## [0.6.0] - 2023-20-09
 ### Changed 
@@ -87,6 +96,7 @@ Internal modifications to SDP cone implementation to reduce allocations.
 
 - Initial release
 
+[0.7.0]: https://github.com/oxfordcontrol/Clarabel.jl/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/oxfordcontrol/Clarabel.jl/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/oxfordcontrol/Clarabel.jl/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/oxfordcontrol/Clarabel.jl/compare/v0.4.1...v0.5.0